BuildToolChain: use ised with .pc file updates too
git-svn-id: svn://localhost/svn/toolchain/trunk@4688 2adf4935-75e9-4407-93b2-34b21a1880a4

+++ buildscript/lib/funcs.sh
@@ -268,6 +268,7 @@ get_scm_ver()
 ised()
 {
 	# not everyone supports the -i option, or does so correctly, so fake it
+	[ -e "$1" ] || return 0
 	sed "$@" > "$1".tmp
 	mv "$1".tmp "$1"
 }
